+++ GridKit/Model/PowerElectronics/InductionMotor/InductionMotor.hpp
@@ -34,7 +34,6 @@ namespace GridKit
     using CircuitComponent<ScalarT, IdxT>::ypB_;
     using CircuitComponent<ScalarT, IdxT>::fB_;
     using CircuitComponent<ScalarT, IdxT>::gB_;
-    using CircuitComponent<ScalarT, IdxT>::jac_;
     using CircuitComponent<ScalarT, IdxT>::param_;
     using CircuitComponent<ScalarT, IdxT>::idc_;
 
@@ -46,7 +45,6 @@ namespace GridKit
     InductionMotor(IdxT id, RealT Lls, RealT Rs, RealT Llr, RealT Rr, RealT Lms, RealT RJ, RealT P);
     virtual ~InductionMotor();
 
-    int allocate();
     int initialize();
     int tagDifferentiable();
     int evaluateResidual();